requirements.txt
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
cryptography==2.8
Flask==1.1.1
Flask-SSLify==0.1.5
gunicorn==20.0.4
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
pycparser==2.19
PyJWT==1.7.1
python-dotenv==0.11.0
requests==2.23.0
six==1.14.0
urllib3==1.25.8
Werkzeug==1.0.0
